+++ src/structs/id_icon_name.cr
@@ -6,7 +6,7 @@ module XIVAPI
     struct IDIconName
       # define a JSON mapping to create instances of this struct
       JSON.mapping(
-        id: {type: Int32?, key: "ID"},
+        id: {type: UInt64?, key: "ID"},
         icon: {type: String?, key: "Icon"},
         name: {type: String?, key: "Name"},
       )
